Log duplicate effect name stacks.
Some effects can have duplicate names.  For example, Harnessed Shadows (Consular Shadow, Infiltrator) stacks 3 times.  When it stacks to 3, a second, invisible effect with the same name is applied to the player, causing this method to fail.  This method logs both of the effects and their effect number.  The effect number can then be used by the rotation to use the correct buff.