Changed Infiltration SingleTarget and AreaOfEffect to lazy-loaded Decorators
Method call wouldn't work.  Should have known that.  The rotation property is only called once and saved by the bot instead of each time it needs it. So we have to make the MainAttack property switch as a result of a delegate call.  Let's just make it the decorator for the composite.